+++ src/app/admin/anime/page.tsx
@@ -2,95 +2,8 @@
 
 import { useState } from 'react';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-
-let animeData = [
-  { id: 1, title: 'Naruto', genre: 'Action, Adventure', rating: 4.5 },
-  { id: 2, title: 'Death Note', genre: 'Mystery, Thriller', rating: 4.7 },
-  { id: 3, title: 'One Piece', genre: 'Action, Adventure', rating: 4.8 },
-  {
-    id: 4,
-    title: 'Attack on Titan',
-    genre: 'Action, Dark Fantasy',
-    rating: 4.9,
-  },
-  { id: 5, title: 'My Hero Academia', genre: 'Superhero, Action', rating: 4.6 },
-  {
-    id: 6,
-    title: 'Fullmetal Alchemist: Brotherhood',
-    genre: 'Adventure, Fantasy',
-    rating: 4.9,
-  },
-  { id: 7, title: 'Demon Slayer', genre: 'Action, Supernatural', rating: 4.8 },
-  { id: 8, title: 'Steins;Gate', genre: 'Sci-Fi, Thriller', rating: 4.9 },
-  { id: 9, title: 'Hunter x Hunter', genre: 'Adventure, Fantasy', rating: 4.9 },
-  { id: 10, title: 'Code Geass', genre: 'Mecha, Thriller', rating: 4.8 },
-  {
-    id: 11,
-    title: 'Dragon Ball Z',
-    genre: 'Action, Supernatural',
-    rating: 4.7,
-  },
-  {
-    id: 12,
-    title: 'Neon Genesis Evangelion',
-    genre: 'Mecha, Psychological',
-    rating: 4.5,
-  },
-  {
-    id: 13,
-    title: 'Cowboy Bebop',
-    genre: 'Space Western, Sci-Fi',
-    rating: 4.9,
-  },
-  {
-    id: 14,
-    title: 'Sword Art Online',
-    genre: 'Adventure, Sci-Fi',
-    rating: 4.3,
-  },
-  {
-    id: 15,
-    title: 'Tokyo Ghoul',
-    genre: 'Dark Fantasy, Supernatural',
-    rating: 4.4,
-  },
-];
-
-const fetchAnime = async (page = 1, search = '') => {
-  await new Promise((resolve) => setTimeout(resolve, 1000));
-  const filteredAnime = animeData.filter(
-    (anime) =>
-      anime.title.toLowerCase().includes(search.toLowerCase()) ||
-      anime.genre.toLowerCase().includes(search.toLowerCase())
-  );
-  const itemsPerPage = 10;
-  const paginatedAnime = filteredAnime.slice(
-    (page - 1) * itemsPerPage,
-    page * itemsPerPage
-  );
-  return {
-    anime: paginatedAnime,
-    totalPages: Math.ceil(filteredAnime.length / itemsPerPage),
-  };
-};
-
-const deleteAnime = async (id: number) => {
-  await new Promise((resolve) => setTimeout(resolve, 1000));
-  animeData = animeData.filter((anime) => anime.id !== id);
-  console.log(`Anime ${id} deleted`);
-};
-
-const addAnime = async (newAnime: {
-  title: string;
-  genre: string;
-  rating: number;
-}) => {
-  await new Promise((resolve) => setTimeout(resolve, 1000));
-  const newId = Math.max(...animeData.map((a) => a.id)) + 1;
-  const addedAnime = { id: newId, ...newAnime };
-  animeData.push(addedAnime);
-  return addedAnime;
-};
+import { deleteAdminAnime, postAdminAnime } from '@/action/admin/anime';
+import adminQuery from '@/hook/query/admin';
 
 const AnimePage = () => {
   const [animePage, setAnimePage] = useState(1);
@@ -99,20 +12,21 @@ const AnimePage = () => {
   const [newAnime, setNewAnime] = useState({ title: '', genre: '', rating: 0 });
   const queryClient = useQueryClient();
 
-  const { data: animeData, isLoading: isAnimeLoading } = useQuery({
-    queryKey: ['anime', animePage, animeSearch],
-    queryFn: () => fetchAnime(animePage, animeSearch),
-  });
+  const { data: animeData, isLoading: isAnimeLoading } = useQuery(
+    adminQuery.animeList(animePage, animeSearch)
+  );
 
   const deleteAnimeMutation = useMutation({
-    mutationFn: deleteAnime,
+    mutationFn: deleteAdminAnime,
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['anime'] });
+      queryClient.invalidateQueries({
+        queryKey: adminQuery.animeList().queryKey,
+      });
     },
   });
 
   const addAnimeMutation = useMutation({
-    mutationFn: addAnime,
+    mutationFn: postAdminAnime,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['anime'] });
       setIsAddAnimeModalOpen(false);
