+++ src/hook/query/index.ts
@@ -1,0 +1,7 @@
+import adminQuery from './admin';
+
+const queries = {
+  adminQuery,
+};
+
+export default queries;
